Backend/requirements.txt
pymongo==4.16.0
dnspython==2.8.0
python-dotenv==1.2.2
fastapi==0.135.1
python-multipart==0.0.20
uvicorn==0.41.0
bcrypt==4.3.0
PyJWT==2.10.1
pyotp==2.9.0
